Shapes are also labels now.
The prefix before the unit is now called the shape.
Label names can be used as units to create new labels. The syntax for this requires @.
Label assignment is also declaration now.
What would be considered a shape before is now called a default label.
A variable is compatible with any label of same stack size.

Also began adding support for 64-bit